requirements.txt
opencv-python>=4.8.0
mtcnn>=0.1.1
mss>=9.0.1
APScheduler>=3.10.4
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
tensorflow>=2.13.0
PyQt5>=5.15.0
plyer>=2.1.0
psutil>=5.9.0
pyinstaller>=5.13.0
auto-py-to-exe>=2.40.0